Add Linux build support for openSYDE, CAN Monitor, and SYDEflash
Port all three GUI applications to build and run on Linux with minimal
changes to the existing Windows codebase. Key changes:

Build infrastructure:
- Add toolchain_linux.cmake and build.sh from dev branch
- Add platform conditionals to all three CMakeLists.txt (RC files,
  static plugins, link libraries, Windows driver skip flags)
- Wire Linux target layer into core (tgl_linux, socket CAN, Linux sockets)
- Build vector_blf from source as static lib for CAN Monitor
- Use system Qt6, OpenSSL, zlib, qcustomplot, and gettext on Linux

Platform abstraction:
- Add C_OscIpDispatcherPlatform.hpp to alias WinSock/LinuxSock dispatchers
- Add DLL_Open/DLL_Close/CAN_Status/SetDLLName compat stubs to Linux C_Can
- Guard windows.h/winsock2.h includes in precompiled headers
- Guard Windows APIs in C_HeHandler, C_ImpUtil, C_GiSvPc, C_Uti

Type portability (Linux 64-bit):
- Fix uint64_t/int64_t -> QVariant ambiguous conversions using
  QVariant::fromValue<quint64/qint64>() and static_cast<quint64/qint64>()
- Fix unsigned long long literal ambiguity in h_GetValueAsHex calls

Runtime:
- CAN Monitor uses SocketCAN on Linux (DLL path field = interface name)
- Read CAN bitrate from ip command for bus load calculation
- Stub osy_git_data_model_monitor trigger checker (Windows-only library)

+++ opensyde_tool/src/system_definition/node_edit/datapools/C_SdNdeDpContentUtil.cpp
@@ -1362,24 +1362,24 @@ int32_t C_SdNdeDpContentUtil::h_SimpleConvertToVariant(const C_OscNodeDataPoolCo
       case C_OscNodeDataPoolContent::eSINT16:
          orc_Variant = orc_Content.GetValueS16();
          break;
-      case C_OscNodeDataPoolContent::eSINT32:
-         orc_Variant = static_cast<int64_t>(orc_Content.GetValueS32());
-         break;
-      case C_OscNodeDataPoolContent::eSINT64:
-         orc_Variant = orc_Content.GetValueS64();
-         break;
-      case C_OscNodeDataPoolContent::eUINT8:
-         orc_Variant = static_cast<int32_t>(orc_Content.GetValueU8());
-         break;
-      case C_OscNodeDataPoolContent::eUINT16:
-         orc_Variant = orc_Content.GetValueU16();
-         break;
-      case C_OscNodeDataPoolContent::eUINT32:
-         orc_Variant = static_cast<uint64_t>(orc_Content.GetValueU32());
-         break;
-      case C_OscNodeDataPoolContent::eUINT64:
-         orc_Variant = orc_Content.GetValueU64();
-         break;
+       case C_OscNodeDataPoolContent::eSINT32:
+          orc_Variant = static_cast<qint64>(orc_Content.GetValueS32());
+          break;
+       case C_OscNodeDataPoolContent::eSINT64:
+          orc_Variant = static_cast<qint64>(orc_Content.GetValueS64());
+          break;
+       case C_OscNodeDataPoolContent::eUINT8:
+          orc_Variant = static_cast<int32_t>(orc_Content.GetValueU8());
+          break;
+       case C_OscNodeDataPoolContent::eUINT16:
+          orc_Variant = orc_Content.GetValueU16();
+          break;
+       case C_OscNodeDataPoolContent::eUINT32:
+          orc_Variant = static_cast<quint64>(orc_Content.GetValueU32());
+          break;
+       case C_OscNodeDataPoolContent::eUINT64:
+          orc_Variant = static_cast<quint64>(orc_Content.GetValueU64());
+          break;
       case C_OscNodeDataPoolContent::eFLOAT32:
          orc_Variant = orc_Content.GetValueF32();
          break;
@@ -2213,31 +2213,31 @@ QVariant C_SdNdeDpContentUtil::h_ConvertContentToGeneric(const C_OscNodeDataPool
    {
       switch (orc_Input.GetType())
       {
-      case C_OscNodeDataPoolContent::E_Type::eUINT8:
-         c_Retval = static_cast<uint64_t>(orc_Input.GetValueU8());
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eUINT16:
-         c_Retval = static_cast<uint64_t>(orc_Input.GetValueU16());
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eUINT32:
-         c_Retval = static_cast<uint64_t>(orc_Input.GetValueU32());
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eUINT64:
-         c_Retval = orc_Input.GetValueU64();
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eSINT8:
-         c_Retval = static_cast<int64_t>(orc_Input.GetValueS8());
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eSINT16:
-         c_Retval = static_cast<int64_t>(orc_Input.GetValueS16());
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eSINT32:
-         c_Retval = static_cast<int64_t>(orc_Input.GetValueS32());
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eSINT64:
-         c_Retval = orc_Input.GetValueS64();
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eFLOAT32:
+       case C_OscNodeDataPoolContent::E_Type::eUINT8:
+          c_Retval = static_cast<quint64>(orc_Input.GetValueU8());
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eUINT16:
+          c_Retval = static_cast<quint64>(orc_Input.GetValueU16());
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eUINT32:
+          c_Retval = static_cast<quint64>(orc_Input.GetValueU32());
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eUINT64:
+          c_Retval = static_cast<quint64>(orc_Input.GetValueU64());
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eSINT8:
+          c_Retval = static_cast<qint64>(orc_Input.GetValueS8());
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eSINT16:
+          c_Retval = static_cast<qint64>(orc_Input.GetValueS16());
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eSINT32:
+          c_Retval = static_cast<qint64>(orc_Input.GetValueS32());
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eSINT64:
+          c_Retval = static_cast<qint64>(orc_Input.GetValueS64());
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eFLOAT32:
          if (os32_Role == static_cast<int32_t>(Qt::ItemDataRole::DisplayRole))
          {
             QString c_Precison = QString::number(orc_Input.GetValueF32(), 'g', 9);
@@ -2270,30 +2270,30 @@ QVariant C_SdNdeDpContentUtil::h_ConvertContentToGeneric(const C_OscNodeDataPool
    {
       switch (orc_Input.GetType())
       {
-      case C_OscNodeDataPoolContent::E_Type::eUINT8:
-         c_Retval = static_cast<uint64_t>(orc_Input.GetValueArrU8Element(oru32_Index));
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eUINT16:
-         c_Retval = static_cast<uint64_t>(orc_Input.GetValueArrU16Element(oru32_Index));
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eUINT32:
-         c_Retval = static_cast<uint64_t>(orc_Input.GetValueArrU32Element(oru32_Index));
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eUINT64:
-         c_Retval = orc_Input.GetValueArrU64Element(oru32_Index);
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eSINT8:
-         c_Retval = static_cast<int64_t>(orc_Input.GetValueArrS8Element(oru32_Index));
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eSINT16:
-         c_Retval = static_cast<int64_t>(orc_Input.GetValueArrS16Element(oru32_Index));
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eSINT32:
-         c_Retval = static_cast<int64_t>(orc_Input.GetValueArrS32Element(oru32_Index));
-         break;
-      case C_OscNodeDataPoolContent::E_Type::eSINT64:
-         c_Retval = orc_Input.GetValueArrS64Element(oru32_Index);
-         break;
+       case C_OscNodeDataPoolContent::E_Type::eUINT8:
+          c_Retval = static_cast<quint64>(orc_Input.GetValueArrU8Element(oru32_Index));
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eUINT16:
+          c_Retval = static_cast<quint64>(orc_Input.GetValueArrU16Element(oru32_Index));
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eUINT32:
+          c_Retval = static_cast<quint64>(orc_Input.GetValueArrU32Element(oru32_Index));
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eUINT64:
+          c_Retval = static_cast<quint64>(orc_Input.GetValueArrU64Element(oru32_Index));
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eSINT8:
+          c_Retval = static_cast<qint64>(orc_Input.GetValueArrS8Element(oru32_Index));
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eSINT16:
+          c_Retval = static_cast<qint64>(orc_Input.GetValueArrS16Element(oru32_Index));
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eSINT32:
+          c_Retval = static_cast<qint64>(orc_Input.GetValueArrS32Element(oru32_Index));
+          break;
+       case C_OscNodeDataPoolContent::E_Type::eSINT64:
+          c_Retval = static_cast<qint64>(orc_Input.GetValueArrS64Element(oru32_Index));
+          break;
       case C_OscNodeDataPoolContent::E_Type::eFLOAT32:
          c_Retval = static_cast<float64_t>(orc_Input.GetValueArrF32Element(oru32_Index));
          break;
@@ -2359,18 +2359,18 @@ QVariant C_SdNdeDpContentUtil::h_ConvertScaledContentToGeneric(const C_OscNodeDa
          }
          else
          {
-            //Int is possible but only if in range
-            if (((f64_ScaledValue < 0.0) &&
-                 (f64_ScaledValue >= static_cast<float64_t>(std::numeric_limits<int64_t>::min()))) &&
-                (f64_ScaledValue <= static_cast<float64_t>(std::numeric_limits<int64_t>::max())))
-            {
-               c_Retval = static_cast<int64_t>(f64_ScaledValue);
-            }
-            else if ((f64_ScaledValue >= static_cast<float64_t>(std::numeric_limits<uint64_t>::min())) &&
-                     (f64_ScaledValue <= static_cast<float64_t>(std::numeric_limits<uint64_t>::max())))
-            {
-               c_Retval = static_cast<uint64_t>(f64_ScaledValue);
-            }
+             //Int is possible but only if in range
+             if (((f64_ScaledValue < 0.0) &&
+                  (f64_ScaledValue >= static_cast<float64_t>(std::numeric_limits<qint64>::min()))) &&
+                 (f64_ScaledValue <= static_cast<float64_t>(std::numeric_limits<qint64>::max())))
+             {
+                c_Retval = static_cast<qint64>(f64_ScaledValue);
+             }
+             else if ((f64_ScaledValue >= static_cast<float64_t>(std::numeric_limits<quint64>::min())) &&
+                      (f64_ScaledValue <= static_cast<float64_t>(std::numeric_limits<quint64>::max())))
+             {
+                c_Retval = static_cast<quint64>(f64_ScaledValue);
+             }
             else
             {
                if (os32_Role == static_cast<int32_t>(Qt::ItemDataRole::DisplayRole))
